+++ db.json
@@ -14,58 +14,103 @@
       "id": 1
     },
     {
-      "name": "aaa",
-      "date": "23.03",
+      "name": "aaaa",
+      "date": "12123",
       "products": [
         {
-          "name": "asd",
-          "quantity": "1",
-          "price": "3.00",
+          "name": "edit",
+          "quantity": "24",
+          "price": "8.20",
+          "status": true
+        }
+      ],
+      "id": 2
+    },
+    {
+      "name": "22222",
+      "date": "22222",
+      "products": [
+        {
+          "name": "a",
+          "quantity": "2",
+          "price": "90.00",
           "status": true
         },
         {
-          "name": "edit sd",
-          "quantity": "46",
+          "name": "asd",
+          "quantity": "2",
+          "price": "90.00",
+          "status": true
+        },
+        {
+          "name": "asdasd",
+          "quantity": "2",
           "price": "8.00",
+          "status": true
+        }
+      ],
+      "id": 3
+    }
+  ],
+  "foods": [
+    {
+      "name": "asd",
+      "description": "asd",
+      "img": "",
+      "productsList": [
+        {
+          "name": "asd",
+          "quantity": "2",
+          "price": "8.00",
+          "status": true
+        },
+        {
+          "name": "asd",
+          "quantity": "2",
+          "price": "8.00",
+          "status": true
+        },
+        {
+          "name": "asd",
+          "quantity": "2",
+          "price": "6.00",
+          "status": true
+        }
+      ],
+      "id": 1
+    },
+    {
+      "name": "asd",
+      "description": "2222",
+      "img": "",
+      "productsList": [
+        {
+          "name": "asd",
+          "quantity": "2",
+          "price": "8.00",
+          "status": true
+        },
+        {
+          "name": "siema",
+          "quantity": "3",
+          "price": "12",
+          "status": true
+        },
+        {
+          "name": "lolz",
+          "quantity": "2",
+          "price": "6.00",
           "status": true
         }
       ],
       "id": 2
     }
   ],
-  "foods": [
-    {
-      "name": "posilek 1",
-      "description": "bla bla bla",
-      "img": "",
-      "productsList": [
-        {
-          "name": "mleko2",
-          "quantity": "1",
-          "price": "6.00",
-          "status": true
-        },
-        {
-          "name": "mleko ",
-          "quantity": "1",
-          "price": "3.00",
-          "status": true
-        },
-        {
-          "name": "xd",
-          "quantity": "3",
-          "price": "12.00",
-          "status": true
-        }
-      ],
-      "id": 1
-    }
-  ],
   "actualList": [
     {
       "list": [
         {
-          "name": "s",
+          "name": "asd",
           "quantity": "2",
           "price": "6.00",
           "status": true
