+++ src/components/Lista/TwojaListaBox.jsx
@@ -10,9 +10,7 @@ import { motion } from "framer-motion"
 */
 const TwojaListaBox = ({ product, index, handleDelete, toggleComplete, setFullList }) => {
 
-
     const [editEnabled, setEditEnabled] = useState(false)
-
 
     function editName() {
         setEditEnabled(true)
@@ -20,7 +18,6 @@ const TwojaListaBox = ({ product, index, handleDelete, toggleComplete, setFullLi
     function saveName() {
         setEditEnabled(false)
     }
-
 
     // function setName(event) {
     //     const newName = event.target.value
@@ -32,7 +29,6 @@ const TwojaListaBox = ({ product, index, handleDelete, toggleComplete, setFullLi
     //     })
     // }
 
-
     const handleChange = (event) => {
         const { name, value } = event.target;
 
@@ -43,11 +39,7 @@ const TwojaListaBox = ({ product, index, handleDelete, toggleComplete, setFullLi
         });
     };
 
-
-
-
     return (
-
 
         <motion.section
             initial={{ opacity: 0, }}
@@ -55,14 +47,13 @@ const TwojaListaBox = ({ product, index, handleDelete, toggleComplete, setFullLi
             exit={{ opacity: 0 }}
             transition={{ duration: 0.5, }}
 
-            key={index} className='flex items-center justify-center gap-2 mt-[2rem] md:gap-8'
+            key={index}
+            className='flex items-center justify-center gap-2 mt-[2rem] md:gap-8'
         >
 
             <BsTrash className='hover:scale-[1.5] transition duration-700' size={25} onClick={() => handleDelete(index)} />
 
             <div className='flex flex-col justify-between items-center w-[90%] bg-[#B7C6FF] py-3 px-4 rounded-xl shadow-[0px_5px_4px_rgba(0,0,0,0.35)] sm:flex-row md:px-8'>
-
-
 
 
                 {!editEnabled ?
@@ -92,15 +83,11 @@ const TwojaListaBox = ({ product, index, handleDelete, toggleComplete, setFullLi
                         <button className='font-bold text-white bg-black  uppercase text-sm border-2 p-1 border-white hover:scale-[1.2] hover:border-red-700 transition duration-500' onClick={editName}> edit</button>
 
                         : <button className='font-bold text-white bg-black uppercase text-sm border-2 p-1 border-white hover:scale-[1.2] hover:border-red-700 transition duration-500 ' onClick={saveName}>save</button>}
-
                 </div>
 
 
-
             </div>
-
         </motion.section>
-
     )
 }
